% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{build_headers}
\alias{build_headers}
\title{Build HTTP headers for API requests}
\usage{
build_headers(api_key = "not-needed")
}
\arguments{
\item{api_key}{Character. API key.}
}
\value{
A named character vector of headers.
}
\description{
Build HTTP headers for API requests
}
\keyword{internal}
