% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{get_scored_lengths}
\alias{get_scored_lengths}
\title{Get scored lengths per label}
\usage{
get_scored_lengths(token_sequences, winning_label)
}
\arguments{
\item{token_sequences}{Named list.}

\item{winning_label}{Character.}
}
\value{
Named integer vector.
}
\description{
Returns the number of tokens scored per label based on the divergence point.
}
\keyword{internal}
