% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{trie_max_branching}
\alias{trie_max_branching}
\title{Get max branching factor of trie}
\usage{
trie_max_branching(trie)
}
\arguments{
\item{trie}{A trie from \code{\link[=label_trie]{label_trie()}}.}
}
\value{
Integer. Max children count at any node.
}
\description{
Returns the maximum number of children at any node. This is the minimum
\code{top_logprobs} K needed to capture all sibling alternatives.
}
\keyword{internal}
